// src/shared/business/auth/session-type.ts
export enum JWT_SESSION {
	TOKEN = "JWT_SESSION",
}

export enum SESSION_TYPE {
	USER_SESSION = "USER_SESSION",
	SBS_CLIENT_SESSION = "SBS_CLIENT_SESSION",
	IWP_USER_SESSION = "IWP_USER_SESSION",
	CIP_USER_SESSION = "CIP_USER_SESSION",
	AGENT_USER_SESSION = "AGENT_USER_SESSION",
	FUND_TRANSFER_SESSION = "FUND_TRANSFER_SESSION",
	PARTNER_SESSION = "PARTNER_SESSION",
	PARTNER_USER_SESSION = "PARTNER_USER_SESSION",
	ECDFCP_USER_SESSION = "ECDFCP_USER_SESSION",
	OLD_TOKEN_SESSION = "OLD_TOKEN_SESSION",
	OCTO_WEB_SESSION = "OCTO_WEB_SESSION",
	OCTO_WEB_ADVANCED_SESSION = "OCTO_WEB_ADVANCED_SESSION",
	AGENT_SESSION = "AGENT_SESSION",
	SIGN_CONTRACT_SESSION = "SIGN_CONTRACT_SESSION",
	OCTO_GUEST_SESSION = "OCTO_GUEST_SESSION",
	OCTO_BASIC_USER_SESSION = "OCTO_BASIC_USER_SESSION",
}